Use size_t for stack-sizes.

+++ cpukit/rtems/src/taskcreate.c
@@ -57,7 +57,7 @@
 rtems_status_code rtems_task_create(
   rtems_name           name,
   rtems_task_priority  initial_priority,
-  uint32_t             stack_size,
+  size_t               stack_size,
   rtems_mode           initial_modes,
   rtems_attribute      attribute_set,
   Objects_Id          *id
